applied long double for better precision

+++ fractol.h
@@ -6,7 +6,7 @@
 /*   By: cherrewi <cherrewi@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/12/27 20:06:50 by cherrewi      #+#    #+#                 */
-/*   Updated: 2023/01/16 20:56:12 by cherrewi      ########   odam.nl         */
+/*   Updated: 2023/01/18 12:06:10 by cherrewi      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,10 +29,10 @@ typedef struct s_image
 
 typedef struct s_grid
 {
-	double	re_min;
-	double	re_max;
-	double	im_min;
-	double	im_max;
+	long double	re_min;
+	long double	re_max;
+	long double	im_min;
+	long double	im_max;
 }	t_grid;
 
 typedef struct s_mlx_data
@@ -46,8 +46,8 @@ typedef struct s_mlx_data
 }	t_mlx_data;
 
 t_mlx_data	define_mlx_window(void);
-t_grid		define_grid(double re_min, double re_max,
-				double im_min, double im_max);
+t_grid		define_grid(long double re_min, long double re_max,
+				long double im_min, long double im_max);
 t_image		define_img(void *mlx_p, int height, int width);
 void		add_pixel(t_image *img, int x, int y, int color);
 void		calc_mandelbrot(t_mlx_data *mlx, int (*f_color)(int));
